[IMP] planning: Remove group to show percentage
Purpose
=======
Currently, the allocated percentage is only
displayed for user in a specific group.
There are no real reason to hide this field to
other users.
Moreover, this group only used once: when
project_forecast is installed to add all planning users
to this group.

Specification
=============
Show the allocated percentage to every user.

Task 2229296