[IMP] timesheet_grid: Remove default groupby from some Timesheets menu
Currently, The My Timesheets, All Timesheets, To Validate > Last Week
and To Validate > Last Month  action has a default groupby on some
field and the timesheet gridview has the same grouping defined in the
view architecture so the default action grouping therefore only makes
sense if we want the other views of the action to be grouped as well,
and we do not want that.

So in this commit we removed  the default_search_groupby* from action
of My Timesheets, All Timesheets, To Validate > Last Week and
To Validate > Last Month.

closes odoo/enterprise#9694

Taskid: 2230634
Closes: #9694
Related: odoo/odoo#49910
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>