[FIX] l10n_es_reports: full and PYMES balance sheets: compute "resultado del ejercicio" on the whole history
This line needs to be computed using the whole accounting history. It shouldn't always reflect the amount of the profit & loss. Unaffected results from previous fiscal years should be manually moved to Remanente account, but if it isn't done, they are expected to stay in this line.

(For the record, this is pretty much what is done in Belgium as well)

closes odoo/enterprise#9869

X-original-commit: 52daff8d487a2c00fcce6a0ffba457ac370de2a5
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>