[REF] social: introduce a manage comment method to ease inheritance
This commit simply introduces a "isCommentManageable" method that is used to
check if we have to display the 3 dots menu that allows managing a comment in
the social Feed view.

This is done in preparation of the social_crm bridge module that will add its
own condition to display that menu.

LINKS

Task ID-2206496
ENT PR odoo/enterprise#20383
UPG PR odoo/upgrade#2763