[MOV] social_demo: reorganize demo data
As we start to have many social media implementations, having a single file to
reference all the demo data was not convenient anymore.

This commit just moves the code around so that we have a "base" file containing
the common data as well as one specific file for every social media
implementation.

Task-2209729
ENT PR odoo/enterprise#14027