[REM] {delivery,stock}_barcode: remove picking form barcode handler
Barcode handling within picking form is outdated and obsolete due to
barcode app. To avoid having to maintain the same barcode features in
both stock (picking form) and barcode apps, we remove the picking
form barcode handler code. It is expected that most users will just use
the barcode app anyways. This includes removing the now obsolete
delivery_barcode module which only added the ability to scan a package
type barcode in a picking form.

Note that other form barcode handlers that exist at the time of this
commit are left as is.

Task: 2039720
Upgrade PR: odoo/upgrade#3230

closes odoo/enterprise#24145

Signed-off-by: Steve Van Essche <svs@odoo.com>