[FIX] l10n_mx_edi: Avoid website errors on sale process
In the next case:
1. Configure in sales the option to auto invoice the orders from the website
2. Configure a payment acquirer
3. Configure the Mexican data in the product to allow generate the CFDI correctly
4. Make a sale from the website and pay it

In this case, without this fix returns the error that the invoice was not signed and
cannot be paid, this for the next 2 reasons:
1. In some cases, the payment is not added in the transaction, this is generated after
and assigned after of invoice generation, then, the `if` that disable de after commit is not
considered and the invoice is not signed because is not processed any commit, now only check
that the invoice has a transaction-related.
2. Now the compute fields are not calculated all times, then when the `l10n_mx_edi_is_required`
method check that the invoices has UUID, this is empty because was not recalculated when the
``l10n_mx_edi_pac_status`` was updated. For this was added in the api.depends and the process
now is correct.

closes odoo/enterprise#9942

X-original-commit: 6abdbc018ddfd040eab31d5d38ec1310990df414
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>