[FIX] account_winbooks_import: Fix runbot errors
Fix various errors that made the tests fail. First, a country needs to
be defined for the test company, as it is a requirement for taxes.
Then, add a further filter for accounts, journals and analytic accounts
on the current company, as otherwise it would mix up accounts and
journals of other existing companies in the database and the resulting
data would be incompatible.

closes odoo/enterprise#9757

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>