[IMP] sign: Add activity type: request signature
Add Activity type: signature. Instead of button 'done' you have
a button to request a signature. When click on it, a modal opens to
choose template and signatories, when click on send, the activity
is done.

On advantage of contract_salary application, you can automatically
create activity when an employee choose this advantage (the value must
!= O or False). This activity can be generated either when the employee
sign or when the hr manager counter sign (once the modification is done
or the advantage is taken).

task ID 2213188

closes odoo/enterprise#9821

Related: odoo/upgrade#1792
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>