[IMP] *,web_studio: Hide default user image for unassigned records
Impacted modules: account_budget, helpdesk, mrp_plm, planning,
quality_control, sale_subscription, web_studio

The default image displayed when the record is unassigned is
confusing, so we are removing it.

Studio is also adapted to reflect this change.
When editing a kanban view, the editing card is
based on the first record.
If the kanban arch already has an avatar but
this first record is not assigned, a placeholder
icon is displayed instead.

FP request

Task 2234524

closes odoo/enterprise#9834

Related: odoo/odoo#49333
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>