[FIX] helpdesk: Filter tickets by "SLA Not Failed" not working properly
The filter was not showing closed tickets, even if they did pass all
their SLA policies. This was due to a wrong search domain returned by
the search attribute of the sla_failed field.

After this commit, the domain returned is correct and thus the filter
works as expected.

Task ID: 2229702

X-original-commit: d4f0cebfa0db238ce2daa80f7e076e43b9347dad