[FIX] website_helpdesk_form: Fix website link in helpdesk team form
Before this commit, website form link that is shown under the
corresponding checkbox in the Helpdesk Team form (if Website Form is checked) was broken, as it
was a relative path but it did not append the base url at the beginning.

After this commit, the link is complete and correctly redirects you to
the corresponding form in the webpage.

Task ID: 2229702

X-original-commit: 0bc43e98ff6d556459e06f601619e27c2922ad86