[FIX] helpdesk: Assignment method changing automatically
Before this commit, the assignment method in the helpdesk team form
changed automatically from Random or Balanced to Manually if the user
removed the last member in Team Members or Team Visibility while the
other was empty. This behaviour didn't add any value, as the user will get a
warning if they try to save the record with the assignment method set to
other than Manually without any Team Members added. And it can be
annoying for the user in case they added the wrong member because of a
typo, and want to remove it and add the correct one, as they will have
to change the assignment method again.

After this commit, the assignment method does not change when there are
no members in Team Member or Team Visibility.

Task ID: 2229702

closes odoo/enterprise#10410

X-original-commit: 2dc3a53ce5387a7ba16291a43f8b75e34e9d7e05
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>