[FIX] l10n_be_hr_payroll_fleet: Data not available at install
Before this commit:
If you have at least one Belgian company with vehicle, it will be
impossible to install l10n_be_hr_payroll_fleet. Indeed as the commit
https://github.com/odoo/enterprise/commit/b56e1facc28b2d784e6dd11046ef167f80165e06
says: "Because Odoo computes stored computed fields before
loading data, the computation will crash for existing records if it's using
a parameter defined in the same module which defines the field (the parameter
does not exist yet)."

After this commit:
The compute field is removed. At creation, if no recurring_cost_amount_depreciated,
we take the default value on vehicle.

closes odoo/enterprise#9974

X-original-commit: 4a741626692b828eee1650c7e5d24f90c718d2b7
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>