[MOV] *_mobile,mail_enterprise: web_mobile.rpc becomes web_mobile.core
This commit renames `web_mobile.rpc` to `web_mobile.core` as it has
nothing to do with 'rpc'... but looks way more like an
equivalent to `web.core` (but for the native/mobile apps).

In summary, it handles the initialization of the native plugins/methods
provided by the mobile apps.

Note: implementation is unchanged.