[FIX] account_followup: exclude selection working on send
Enable 'Join open invoices' on the followup reminder
Have a partner with multiple unpaid invoices
Go to followup reports, select such partner and from the view check some
lines to be excluded.
Send mail

Excluded choices will be ignored and all lines will be sent in the
reminder email. Moreover even if some invoice is excluded and the
attachment has not been generated yet, the whole action will be blocked
by the message "You are trying to send a followup report to a partner for
which you didn't print all the invoices".

Using a more complex filter based on the actual invoice move lines and
reusing the same lines to select which attachments to send solve the
issue

opw-2238836

closes odoo/enterprise#9984

X-original-commit: 24be36872bf09d764052938603778203221d3e05
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>