[FIX] l10n_de_reports: datev export fix 0% taxes
Create an invoice, assign a customer and create a single invoice
line with product and 0% tax.
Go to Accounting>General Ledger and click on "Export Datev (ZIP)".
The file EXTF_accounting_entries.csv into the zip package
report the acccounting lines.

The line related to the invoice will have a 0 amount, because the taxes
are matched based on their tax_line_ids, which a 0% tax does not trigger
since no accounting move are created for 0 tax amounts. Attempting to
fix manually counting the balance since the tax amount would be 0

opw-2230699

closes odoo/enterprise#10170

X-original-commit: 25b71ab6dc55dd1399b564307cf249c1b0b572fc
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>