[IMP] account_sepa: added SE dometic payments
Swedish payments is compliant with ISO20022 but its interpretation differs from SEPA, some fields are not the same: InitgPty, Dbtr, DbtrAgt and Cdtr. A new version of sepa has been added which is really the swedish payment and allows the payment's file generation methods to reflect these small changes.

Task : 2216105

closes odoo/enterprise#11217

X-original-commit: 7663821badaed313abcf3741687742c22eea9b84
Signed-off-by: Josse Colpaert <jco@openerp.com>