[FIX] delivery_easypost: send predefined package and parcel dimensions
Before this commit, if the user didn't use a specific package it's
considered that he put everything inside a single package. But only the
weight of the package was sent to easypost.

Now, the predefined package and the parcel dimensions if exists are also
sent.

Fine-tuning of 6487e447ec8f925c04b72b0a5e355ab689ea7dfd

opw-2209365

closes odoo/enterprise#10135

X-original-commit: d91cc166b405d10044222b2a0f356da4b456b3cc
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>