[IMP] web_enterprise: avoid dropdown when there is one action on Mobile
Before this commit, even for a single visible CTA in statusbar, there
was an 'Action' dropdown, adding an extra click for end user.

After this commit, if there is only one visible CTA, the dropdown is not
displayed anymore on mobile devices.

Steps to reproduce:
* Go in the Mobile App
* Go in Project
* Select a Task with a TimeSheet
* There is a dropdown menu and only one action in.

Task ID: 2225270

closes odoo/enterprise#10602

Related: odoo/odoo#51265
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Co-authored-by: Siddarth Gajjar <sga@odoo.com>