[FIX] account_reports: Allow disable partner_categories on filters
Is possible add the filter_partner in some account reports, like in
partner ledger:
https://github.com/odoo/enterprise/blob/d91cc166b405d10044222b2a0f356da4b456b3cc/account_reports/models/account_partner_ledger.py#L21

Is this is enabled, are added 2 options:
- Partner
- Category

If in a custom report is necessary hide the filter for Category is not
possible, because not exists the option filter_partner_category = False,
then, could be inherited the method
[_get_options](https://github.com/odoo/enterprise/blob/12.0/account_reports/models/account_report.py#L130)
and remove that option, but if its removed I get an error on report
generation because get the values with dict['key'] and not with
dict.get('key').

Without this fix, is necessary overwrite the method
``get_report_informations`` to make this improve and avoid the error
(Its no the correct).

closes odoo/enterprise#10301

X-original-commit: 223c738c78d243aea521a1dbc5a015fab04ff93f
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>