[FIX] hr_appraisal: Filter appraisal report by company
Go to Appraisal > Reportoring > Calendar view
=> Access error: cannot read employee

Appraisal reports are not filtered by enabled company.
Therefore, appraisal of every companies are included in
the report.
However the employee name is read as the display_name. This is
not allowed if the user does not have access to the employee's company.

Two solutions:
1) Filter report records by company: add a company field
   in the SQL view and the associated ir.rule
2) Read the employee name as sudo and display appraisal of
   all companies (related fields are compute_sudo=True by default)

Option 2) is a safer change in a stable release and does not
require a module update.

Task 2243819

closes odoo/enterprise#10265

X-original-commit: 08969dbe9430c6da2d2a723cb141acf3579d46dc
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>