[FIX] industry_fsm_{sale,stock}: Open serial number wizard
Since c79aa9a, fhe FSM/stock integration is broken with product tracked
with serial numbers.
Go to a new FSM task > Products > try to add a tracked product
=> nothing happens.
A wizard to manage the serial number should open
The action returned prior to c79aa9a is simply ignored
since everything is managed in an inverse method which is
not expected to return.

This commit extract part of the business logic out of the
inverse method into new public methods which can return
an action.
The web client is adapted to save this particular field
and react to any action returned when saving the field.

Task 2245645

closes odoo/enterprise#10636

X-original-commit: d9594d86daf974e70c0cb25208adb54ecf3dd9ef
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: lul-odoo <LucasLefevre@users.noreply.github.com>