[FIX] account_reports: Fix default filter_accounts on General Ledger
You can open the General Ledger from the journal dashboard by clicking on
"Balance in GL" for cash/bank journals. The idea was to set a default value for
the accounts filter input.

This didn't work because the filter only works in the current DOM and then,
you wasn't able to remove the content of the filter and see the rest of the report.
Indeed, the lines has been filtered Python-side.

This commit fixes the issue by managing the default filter accounts input value directly
js-side.