[FIX] hr_contract_salary: drop language in context
The call to compute_sheet will (indirectly) create an hr.payslip.line
that whose name is translatable. This means an ir.translation record
will be created.

If an applicant visit the salary with his browser in a language not
activated, the context will contain this language and a translation
will be created (and crash)

...
  File "hr_contract_salary/controllers/main.py", line 424, in compute_net
    payslip.with_context(salary_simulation=True).compute_sheet()
  File "hr_payroll/models/hr_payslip.py", line 166, in compute_sheet
...
  File "odoo/models.py", line 3907, in _create
    Translations._set_ids(tname, 'model', self.env.lang, record.ids, val, val)
  File "odoo/addons/base/models/ir_translation.py", line 279, in _set_ids
    for res_id in set(ids) - set(existing_ids)
...
ValueError: Wrong value for ir.translation.lang: 'nl_BE'

This patch hides the error by dropping the visitor language

closes odoo/enterprise#10267

X-original-commit: d80ad13af1999cc27bea45cf2ef3806fd439bc49
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>