[FIX] planning: compute the allocated_hours for demo data
Currently, Due to default value the allocated_hours is not computed
each time when the plannig slot is created from the demo data. it
is going to set the default value instead of computing the value.

So in this commit remove the default value on the field so it
should compute the value each time and set the value 0.0 in compute
if the compute data is not available accroding to condition.

closes odoo/enterprise#10412

Taskid: 2246537
Closes: #10271
X-original-commit: 27d94e3e249f48648cfedd294f75febfdba9433f
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>