[FIX] l10n_be_hr_contract_salary: Remove customer_relation field
Purpose
=======

The field customer relation is used to allow employees taking
representation fees advantage with a maximum of 150€ or 110€
if they have direct contacts with customers or not.

During the internationalization of the salary configurator, this has
been forgotten.

It has been fixed at https://github.com/odoo/enterprise/pull/10121/commits

But, this value is never set to False in practice. And if one day, we
wish to make an exception, then we can always create a specific contract
template for this employee.

closes odoo/enterprise#9536

Taskid: 2214356
Related: odoo/upgrade#1127
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>