[FIX] documents_sign, sign: default favorite
- Install `documents_sign`
- Set the access rights of the Demo user to 'Documents > User' and
  'Sign > User: Own and Shared Templates'
- Connect as Demo
- Upload a PDF file
- Choose 'Create signature template(s)'

An `AccessError` is raised.

The following record rules are applied:

https://github.com/odoo/enterprise/blob/5b3624540eeadf5fcd19d79f8f3c8bfcfdcf854c/sign/security/security.xml#L22-L36

Since the default `privacy` is `invite` and no `favorited_ids` is set,
the user doesn't have access to the record he is creating.

We set the `favorited_ids` if the `documents_sign` and also add the
default value for the field to prevent further issues.

opw-2240423

closes odoo/enterprise#10357

X-original-commit: 23e9f39b3b1ed1892cd39d07039249716172481f
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>