[REF] documents: allow to override the preview class
Before this commit, the class of the preview was based mainly on
document type and assigned in the template, it was difficult to extend
this behavior.

After this commit the class is computed in code and bound to the template.

It's useful for the documents spreadsheet integration

Part of task-ID 2223145