[IMP] planning: Added action and stat button to view planning on employee
Purpose of the task is, the planning gantt view doesn't allow to search
on employees, but only on slots. This is limiting when you need to
identify the availabilities of a set of employees based on their skills
or default roles.

So in this commit, Added the 'View Planning' on employees action menu
which should redirect to the Planning of the selected employees with
group by employee.
Also added the stat button on employee form view which will redirect to
planning view of particular employee.

When opening the plannning view which is redirected by planning action,
we should display the line for each of the selected employees, even if
this employee doesn't have any shifts planned so did the changed on
read_group_employee_id. also hide the open shift for employee when
opening the gantt view from employee.

closes odoo/enterprise#10404

Taskid: 2250008
Closes: #10404
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>