[FIX] voip: no contact shown on incoming call
Before this commit, for an incoming call, only the phone number was
shown, even if the contact is registered as a contact.

The phone number used to search the contact, has the standard USA
formatting domestic numbers convention '1 (are code) extension', without
international call prefix. The numbers in the contact uses '+' as
international call prefix. For example: 12686152726 call and a contact
using this phone number is searched instead of +12686152726 (which is
the value stored in the db).

Fine-tunning of aabdb279e24be696bcd48d3b79703bdfb2c6c56f

opw-2244047

closes odoo/enterprise#10363

X-original-commit: 8051470d88a803229567d7644cce07379c7c38a8
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>