[FIX] account_asset: deadlock with asset and reset to draft invoice
Usecase:
* Setup an account to generate an asset
* Create an invoice with a line using the account
* Asset is generated automatically, validate it
* Post a depreciation line
* Reset depreciation to draft
* Reset original invoice to draft

Now you cant open open the asset because the _compute_value is
trigerred by the invoice state change, not allowing to have draft
original lines.
And you can't post the invoice again because "One of the selected
Journal Items already has a depreciation associated"
The user is now in a deadlock.

The solution is to allow the user to post the invoice again, without
creating a new asset. The compute function will not be called and the
user can correcly close/recompute the depreciation board/do something
else on the Asset.

Forward Port of https://github.com/odoo/enterprise/pull/9525/

closes odoo/enterprise#10537

X-original-commit: 00894820576a0f4eab6f9573dc5e0a20ce2b68c4
Signed-off-by: Simon Goffin (sig) <sig@openerp.com>
Signed-off-by: wan <william-andre@users.noreply.github.com>