[FIX] helpdesk: dashboard use open_hours not close_hours
In c104926332 close_hours ("Open Time (hours)") that was a stored field
updated with a cron, was replaced by:

- close_hours ("Time to close (hours)") a stored field set when a ticket
  moved to a "closed" stage

- open_hours ("Open Time (hours)") equivalent of original close_hours
  field but that is not stored.

This was not changed in the dashboard, so the statistics of hours would
be over the number to close a ticket for open tickets (so 0 unless you
close a ticket then re-open it).

With this changeset, we use the open_hours instead to get back to the
behavior before 13.0 version.

opw-2244822
closes #10380

closes odoo/enterprise#10392

X-original-commit: f23e5f3499d5cb4430bcd1f7b65e3077f785cd5a
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>