[FIX] l10n_ar_edi: proper set private key for second company
Before this change when the user try to generate the csr file from the config settings it was creating and adding the private key to the user's default company and not to the current logged company.

With this change now will create and set the private key properly in the current logged company.

closes odoo/enterprise#10665

X-original-commit: 32f6a0941e22e10c8c5e17ddc4214acedff895af
Signed-off-by: Josse Colpaert <jco@openerp.com>