[FIX] marketing_automation: process_event does not support batch
In case several traces are triggered at the same time (set_open / set_clicked
...) a call on process_event is done using a recordset with several marketing
traces in it. However that method waits for a singleton recordset.

This commit fixes that behavior. Note that generally traces are triggered one
by one and having multiple traces is a rare corner case.

Task ID 2228927