[ADD] l10n_ar_edi: Always save xml request/response from AFIP
Before, we where saving this information only when the invoice was properly validated in AFIP or has been validates with observations.

Now we are saving this info when the invoice is been rejected by AFIP, which is really good because let us to interpret errores related to electronic invoice validation.

closes odoo/enterprise#10667

X-original-commit: 2e63064ef024ce9e4d5af5e8ca91df7be45743a2
Signed-off-by: Josse Colpaert <jco@openerp.com>