[FIX] social_facebook: fix deprecated 'comment/likes' endpoint …
As of 5th of May 2020, Facebook has decided to deprecate a few endpoints:
https://developers.facebook.com/docs/graph-api/changelog/2020-02-03-endpoint-deprecations

We thought that those changes did not impact us because we don't directly use
any of those endpoints.

However, graphQL allows you to indirectly nest endpoint calls by accessing
sub-fields on an endpoint definition.

For example, if we ask for the following "fields" on the "comments" endpoints:
'likes.limit(0).summary(true)'

It will indirectly call the endpoint comments/likes, which is part of the
deprecated ones.

This was simply fixed by requesting the "like_count" field instead, which
allows to keep the same behavior as before.
(In other words, nothing changes functionally)

Task 2191695

closes odoo/enterprise#10474

X-original-commit: a08de1c98452606d55a381896c5dd2aff0c32fd7
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>