[FIX] social: fix feed view popover
Bug
===

In social feed view, you can hover the statistics of the accounts to get
more details about the number. The popover is sometimes stuck and user has
to refresh the page to make it disappear.

How to reproduce ?
1. Hover the statistics to make the popover visible
2. With chrome, press "AltGr + LeftArrow" to go back without moving the
   mouse
3. When you'll change the view, the popover will be stuck

Technical
=========

The DOM can be updated in many ways (when we refresh the posts/stats,
with keyboard shortcut...). When the statistics are refreshed, the
buttons are destroyed and new one are created. If it happen when the
popover is opened, it will be stuck because the event "mouse leave"
of the removed button can not be triggered anymore to hide the popover.

Solution
========

The bootstrap popover must be closed before the element is detached from
the DOM (or before the element is replaced by an other one).

Task-2248204

closes odoo/enterprise#12783

X-original-commit: 304dde81c95d88193db51d768d84a7a377b3e65c
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>