[FIX] web_map: still show osm map in case of fail geocoding
To get the location of partner, we either use:

- http://nominatim.openstreetmap.org that is very restricted
  (see https://operations.osmfoundation.org/policies/nominatim/)

- https://www.mapbox.com/ if the user has added a MapBox token in
  "Settings" > "Map Routes".

It seems it is very easy currently to get banned on the first service
(eg. when testing with 36 geocoding request, I was blocked for request
12-36 and 1-11 got through), but when you get banned:

- the queries that got throught are lost (we only save if all worked)
- we do not show the world map

So currently in this situation only MapBox should be working.

With this changeset, when we user is banned of openstreetmap, he will
still see the map and the location queries that worked will still be
saved.

opw-2243126
closes #10469

closes odoo/enterprise#10521

X-original-commit: 09afec4dafce91c6426fc1abcb8429aea9a43afe
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>