[FIX] web_mobile: properly disable datepicker when using native method
When the native datetime picker is used (on mobile apps Android / iOS),
the JS datetime picker is destroyed, resulting into the wrong format
being used when setting the field's value.

The reason behind this is that we detroyed the JS datetime picker and
then hand over the input handling to the native method... or so we
thought! Actually the JS datetime picker is still lurking in the shadows
but, as we destroyed it, its default configuration is applied on the
value once send from the native datepicker, resulting into the wrong
format being applied.

This commit fixes it by not initialising the JS datetimepicker at all
and properly overriding the `web.datepicker` methods that may call the
JS library.

opw-2242880

closes odoo/enterprise#11654

X-original-commit: e231d78ffdb660d1fcf44e6f2c14df1e4a19730c
Related: odoo/odoo#54068
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>