[FIX] hr_payroll_expense: Fix the traceback when wrong access rights
Currently, There is traceback when trying to open an expense report if
the user has the invoicing > advisor or the invoicing > consolidation
user access right levels, but no access rights in expense.

Issue is procuded because the refund_in_payslip has group for expense
team approver but this field is used in attrs of buttons which are
visible by the account manager. so when user has right for account
manager but not expense approver then it will generate the traceback.

So Fix the issue and allow to use the field to account manager so field
should be used in attrs of buttons.

closes odoo/enterprise#10751

Taskid: 2246587
Closes: #10492
X-original-commit: 6577f82f62b8d4c0351cf72846f88a6ef28e5b06
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>