[FIX] timesheet_grid: Fix missing view at module uninstallation
Purpose
=======

Unfortunately, the grid view is defined in enterprise, and the
timesheet actions (community) are inherited in enterprise to
add the grid view in the view_modes.

As they override view_mode directly instead of creating
ir.actions.act_window.view that would be unlinked properly
when uninstalling timesheet_grid, here we clean the view_mode
manually.

But in master, define ir.actions.act_window.view instead,
so that they are removed with the module installation.

closes odoo/enterprise#10519

Taskid: 2186631
X-original-commit: 32e4ba0333da2acd8f490dc88b665133a08a8cb2
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>