[FIX] web_mobile: support --no-database-list in mobile routes tests
When Python tests are run with the `--no-database-list` command line
option the `MobileRoutesTest#test_database_list` test fails because the
route tested is explicitly disabled. Also this option is used on all
processes on Odoo.sh, preventing any build from succeeding.

This commit adds support for the `--no-database-list` command line
option when running this test by adapting it to test both scenario.

opw-2252357

closes odoo/enterprise#10517

X-original-commit: a03b11061da511f02cc4b408926b3755ea71163c
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>