[FIX] web_studio: duplicated tree_view_ref and no default
In studio, when we edit a sub x2many view that is not embedded, we
duplicate the default view and embed it.

This work fine by default, but it can be odd in case the particular view
is using eg. a specific view thanks to a tree_view_ref context key on the
x2many field, then editing it could totally change the content since we
duplicate the default view and not the real view that was being
displayed.

With this changeset, we instead duplicate tree_view_ref if it is
available.

Without change, added test fails with:

  context tree_view_ref should be propagated for inline view creation
  [actual: "undefined", expected: "module.tree_view_ref"]

opw-2251643
closes #10508

closes odoo/enterprise#10678

X-original-commit: fcf5c26810786623fa6b9821b16b16e1e870fb6d
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>