[FIX] sale_stock_renting: scheduling view consistence and performance
1) The scheduling view has shown a performance bottleneck on database with
high numbers of serial numbers.

After analysis of the SQL view definition, a join in the subquery can be
removed to simplify queries and improve performance.

opw-2229482

2) Furthermore, since a recent fix commit https://github.com/odoo/enterprise/commit/ee1385a86db22f12444d871cbb2df591aa97afea ,
serial numbers pickedup but not reserved weren't visible on the gantt view anymore.

This problem is also fixed by the refacoring of the WITH subquery.

X-original-commit: f39ed69f9abb7cb13a2b59778278b01b44b26661