[FIX] sale_stock_renting: scheduling view id generation
As sales order line are duplicated by serial number rented, the id
generation of the scheduling sql view is a concatenation of
lot_id(max_id)soline_id but the max id only considered lots id and sales
orders id instead of sales order line ids.

This means that a lot rented two times in same SO but different lines
could be fused in one record instead of appearing twice on the
scheduling view.

closes odoo/enterprise#10649

X-original-commit: e32628ce6d9bbe26f3d1e66e0f4c8ddb26aeb786
Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>