[FIX] l10n_ar_edi: use clean CUIT number for AFIP
Create certificate request and webservice token always with the vat
number of type CUIT, if not will raise and error.

Since now the l10n_ar_vat is the VAT number cleaning all the dashed and
formats we are avoid having problemas with company configuration and
AFIP configuration.

closes odoo/enterprise#10925

X-original-commit: 87876972a374e67eaf236ef521ee64b4f2602b1d
Related: odoo/odoo#52380
Signed-off-by: Josse Colpaert <jco@openerp.com>