[FIX] web_studio: fix import of studio customizations - Activate Studio from dashboard of database A - Export studio customizations - Activate Studio from dashboard of database B - Import studio customizations previously exported The imported data don't have the boolean studio field set to True.
Since v12, when loading a module, a query is executed to create or update XML ids.
The create and write methods are not called anymore.
Therefore the overridden create and write methods of "ir.model.data" defined in Studio module
are not called and thus cannot set the studio field to True.

The generation of the query creating XML ids has been moved to a private method
allowing another module like Studio to override it.

opw-2245578

closes odoo/enterprise#10586

X-original-commit: 4d75777446883de92eeecd4c9ec00f5f4bf50f23
Related: odoo/odoo#51203
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>