[FIX] mail_mobile: wrong channel type
Before this commit, channels in the message were always empty. So the
type of channel doesn't match the condition and the notifications
in mobile are not formatted and associated correctly.

After this commit, we use the self object as it's the 'mail.channel'
model itself.

Steps to reproduce:
* Activate OCN
* Use a mobile and login at least once on Odoo (e.g. using Demo)
* Then send a direct message (e.g. to Demo) with another user (e.g. Admin)
=> Now on mobile a notification appears but not formatted correctly and
not in the right Android Channel

Note:
I guess the bug was introduced by this commit odoo/odoo@be2795513644f09f6b443f462bb67cc5a0762e8d
when some code was refactored form mail_message to mail_thread.

Task ID: 2229735

Forward-port-of: #9604
X-original-commit: 708e4ddc546649986f91fda9ef97a94a700ad026