[FIX] web_map: update MapBox to use new non-legacy API
MapBox's Legacy Static Images API seems to be soon to be retired:
https://blog.mapbox.com/deprecating-studio-classic-styles-d8892ac38cb4

So eg. for new customer it seems that the tiles would not show, and
after first June the API is soon to be totally retired.

So this commit follow advices for migrating to use
"Mapbox Static Tiles API" instead given in:
https://docs.mapbox.com/help/troubleshooting/migrate-legacy-static-tiles-api/

opw-2243126
closes #10570

closes odoo/enterprise#10674

X-original-commit: 21274c15e81390d2fc4fc1d91f6708ad8335297f
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>