[FIX] l10n_ar_edi: convert onchanges in constraints
Before we were validating the afip certificate and privatekey using and onchange, and after the first onchange warning message popup the res.config.settings view broke.

Now with this change we do the files validation only when saving the wizard and this way we avoid the problem of the broke view.

closes odoo/enterprise#10885

X-original-commit: d2a72558f8f11c6d38d22799b2f2a1b55dad1de4
Signed-off-by: Josse Colpaert <jco@openerp.com>