[FIX] helpdesk_timesheet: Fix error when timesheeting on invoiced ticket
Currently when Enable 'Time Reinvoicing' on your helpdesk team
and creating a ticket in this team -> set the task field -> log time
-> create an SO -> create an invoice -> post the entries.
then go back to the ticket and launch the timer while saving the the
timesheet instead of creating the new timesheet it was going to
generate the validation error like 'You can not modify already
invoiced timesheets'.
Issue is produced because it was going to update the all timesheet
instead of just linking the newly created timesheet.

So in this commit, fix the issue and only link the newly created
timesheet with ticket.

closes odoo/enterprise#10612

Taskid: 2245931
Closes: #10612
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>