[IMP] helpdesk,mrp_plm: Remove the message_needaction_counter from kanban view
Purpose of the task is to remove the needaction message counter from
kanban view due to below reason:
 - The user has access to unread messages on records from the discuss
item in the systray.
 - It is irrelevant for user handling notifications through their mail
client
 - It is only less UI item to worry about on the kanban views (as
features increase, we need to be careful not to clutter the views)
 - The user can still use the message_needaction filter to get records
with unread messages

So in this commit we removed the need action counter from all the
kanban view.

closes odoo/enterprise#10613

Taskid: 2257624
Closes: #10613
Related: odoo/odoo#51299
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>