[FIX] voip: no contact shown on incoming call
Since 8c03446f57999ed23d9ee9e77797b22af9606063, to find the contact of
an incoming call, a sanitized process is done to add the international
code if is missing. This relay in the inviteSession.remoteIdentity.uri.type
parameter, which is not always set.

Now a fallback is done, it checks if the last 6 digits of the phone
number corresponds to the last 6 digits of a phone number in the
contact. Like this we can find the contact without having the
international code.

opw-2254068

closes odoo/enterprise#10635

X-original-commit: 1cae30a425161da0dba7e0ea94abf8b6cc23e38c
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>