[IMP] various : Update the demo data of SO/PO/invoice with activities
Purpose of the task is, we have added the activities on various
listviews, the goal of this task is to update SO/PO/invoice demo data
with activities and also set tags on sales orders.

So in this commit, We added the activity on several SO/PO/Invoices and
also added the some tags one sales order and set the user as False
where the user is set as 'Odoobot'.

closes odoo/enterprise#10743

Taskid: 2254708
Closes: #10743
Related: odoo/odoo#51799
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>