[FIX] planning: fix planning template offset - Go to Planning - Go to Configuration > Roles - Create a new role ("TEST_ROLE") - Go to Configuration > Shift Templates - Create a new shift template with the following data:   * Role: "TEST_ROLE"   * Start hour: 06:20   * Duration (hours): 7:24 - Go to My Planning > Calendar and switch to gantt view - Add a new shift - In the wizard select the Planning Template created for "TEST_ROLE".   The options must look like "6:20 AM - 1:44 PM TEST_ROLE" Once the Planning Template selected, the Start date and End date are automatically set, but there are offset by one minute compared to the time configured in the template.
An onchange is executed when template is selected to compute Start and End date.
The onchange method performs float operations to compute the Start date that result on
a one minute difference.

As the time is diplayed correctly for the name of the template, the same operations are used
to retrieve the hours and minutes of the Start date.

opw-2253554

closes odoo/enterprise#10664

X-original-commit: 334754d249218fdf8144ac2afad9fb97fa228686
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>