[FIX] account_reports: correct sign for cash basis taxes
The report directly considered the journal in which each move was done in order to influence the sign modificator to apply to the move lines. While correct for other cases, this was wrong for cash basis moves, generated by taxes based on payment (aka cash basis taxes). For them, the journal type to be considered is the one in which their original invoice had been done.

closes odoo/enterprise#10857

X-original-commit: dd0a19af1cfe1e8f8278a4aee8f1272bc5b974cf
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>